Refactor
Pulled out the meta data extraction code and moved it to the [meta data repo](https://github.com/ch3njust1n/conference_metadata), started on ICML and refactored code into conference object hierarchy.
