+++ config.ini
@@ -1,6 +1,6 @@
 [DEFAULT]
 conference = neurips
-year = 2020
+year = 2018
 title_kw = graph
 author_kw = 
 affiliation_kw =
